+++ src/polymer-build.ts
@@ -5,6 +5,7 @@ export {BuildBundler} from './bundle';
 export {addCustomElementsEs5Adapter} from './custom-elements-es5-adapter';
 export {forkStream} from './fork-stream';
 export {HtmlSplitter} from './html-splitter';
+export {htmlTransform} from './html-transform';
 export {jsTransform} from './js-transform';
 export {getOptimizeStreams, OptimizeOptions} from './optimize-streams';
 export {PolymerProject} from './polymer-project';